requirements.txt
# 量化交易系统依赖包
# =====================

# Web框架
Flask==2.3.3
Flask-CORS==4.0.0

# HTTP客户端
requests==2.31.0

# 数据处理
pandas==2.0.3
numpy==1.24.3

# 数据库
SQLAlchemy==2.0.19

# 异步编程
asyncio-mqtt==0.11.1
websockets==11.0.3

# 数据可视化
matplotlib==3.7.2
plotly==5.15.0

# 技术分析
TA-Lib==0.4.27

# 配置管理
python-dotenv==1.0.0
PyYAML==6.0.1

# 日期时间处理
python-dateutil==2.8.2

# 加密
cryptography==41.0.3

# 测试框架
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 代码质量
flake8==6.0.0
black==23.7.0

# 开发工具
python-decouple==3.8

# 实时数据
redis==4.6.0

# 日志处理
loguru==0.7.0

# JSON处理
ujson==5.8.0

# 时区处理
pytz==2023.3

# 网络请求
httpx==0.24.1

# 类型检查
mypy==1.5.1

# 文档生成
Sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# API文档
flask-restx==1.1.0

# 数据验证
marshmallow==3.20.1

# 缓存
Flask-Caching==2.0.2

# 会话管理
Flask-Session==0.5.0

# 安全
Flask-Talisman==1.1.0

# 限流
Flask-Limiter==3.3.1

# 监控
prometheus-client==0.17.1

# 任务队列
celery==5.3.1

# 数据库迁移
Flask-Migrate==4.0.4

# 文件处理
openpyxl==3.1.2
xlsxwriter==3.1.2

# 邮件发送
Flask-Mail==0.9.1

# 数据序列化
pickle5==0.0.12

# 时间处理
arrow==1.2.3

# 系统监控
psutil==5.9.5

# 文件监控
watchdog==3.0.0

# HTTP服务器
gunicorn==21.2.0
eventlet==0.33.3

# 开发服务器
Werkzeug==2.3.6